move some core components that got missed in the first round
