+++ frontend/src/app/[locale]/(base)/grantor/opportunity/[id]/edit/_components/OpportunityAttachmentUploadInput.tsx
@@ -15,7 +15,7 @@ import {
   ModalToggleButton,
 } from "@trussworks/react-uswds";
 
-import { DeleteAttachmentModal } from "src/components/application/attachments/DeleteAttachmentModal";
+import { DeleteAttachmentModal } from "src/components/core/fileInput/DeleteAttachmentModal";
 import { USWDSIcon } from "src/components/core/USWDSIcon";
 
 const MAX_FILE_SIZE_BYTES = 2 * 1024 * 1024 * 1024; // 2GB
