+++ frontend/src/components/core/IconInfoSection.tsx
